Fix false parallel parameter value to a string
The boolean parameter `parallel`[1] does not work as expected in a CircleCI
job. Per CircleCI docs[2], a boolean field may be converted to 0/1, which is
what appears to be happening in this case.

Given `parallel: false` (or unset, as in the example simple config[3]), the
environment variable `COVERALLS_PARALLEL` is set to `0`, which fails the test
in the coverage reporter[4].

This changeset forces the value 0 back to the string "false" to satisfy the
test in the coverage reporter.

1. https://github.com/coverallsapp/orb/blob/f9b631db7ef9e8ca4fc0e79422c2d3882435ec5b/src/%40orb.yml#L109
2. https://circleci.com/docs/reusing-config/#boolean
3. https://circleci.com/developer/orbs/orb/coveralls/coveralls?version=2.1.0#usage-simple
4. https://github.com/coverallsapp/coverage-reporter/blob/8a6ea39858c9e659839d73fad302540c79740570/src/coverage_reporter/cli/cmd.cr#L91