Ensure that `coveralls` command succeeds when fail-on-error is false
Previously, if the `coveralls.sh` script fails to download the coverage
reporter, running `coveralls report` resulted in a command-not-found
error (error code 127). This failure once occurred multiple times over
the course of a day for me, so it's not not just a theoretical concern.

Fixes https://github.com/coverallsapp/orb/issues/40.